랜덤마라톤 week30 B번

+++ solvedac/random_marathon/week30/11880_개미.kt
@@ -6,7 +6,7 @@ import java.io.InputStreamReader
 fun main() {
     val br = BufferedReader(InputStreamReader(System.`in`))
     repeat(br.readLine().toInt()) {
-        val (a, b, c) = br.readLine().split(" ").map { it.toInt() }
+        val (a, b, c) = br.readLine().split(" ").map { it.toLong() }
         val caseA = (pow2(a) + pow2(b + c))
         val caseB = (pow2(b) + pow2(c + a))
         val caseC = (pow2(c) + pow2(a + b))
@@ -14,4 +14,4 @@ fun main() {
     }
 }
 
-fun pow2(x: Int): Int = x * x
+fun pow2(x: Long): Long = x * x
